requirements.txt
alembic==1.6.5
attrs==21.2.0
autopep8==1.5.7
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.0
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
markdown2==2.4.0
MarkupSafe==2.0.1
mock==4.0.3
moviepy==1.0.3
numpy==1.21.0
packaging==20.9
Pillow==8.3.0
pluggy==0.13.1
proglog==0.1.9
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.13.3
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
tqdm==4.61.1
urllib3==1.26.6
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
